Lecture12/requirements.txt
python-dotenv
joblib
pandas
rich
agno